requirements.txt
absl-py==1.0.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.8
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==5.0.1
atom3d==0.2.6
attrs==24.2.0
biopython==1.79
blosc2==2.5.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
conda-pack==0.8.1
contextlib2==21.6.0
contourpy==1.3.0
cycler==0.12.1
dill==0.3.9
dm-haiku==0.0.9
dm-tree==0.1.6
easy-parallel==0.1.6
filelock==3.17.0
flatbuffers==24.3.25
fonttools==4.55.3
freesasa==2.2.1
frozenlist==1.5.0
fsspec==2024.10.0
gast==0.4.0
google-auth==2.36.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.68.0
h5py==3.12.1
huggingface-hub==0.29.2
idna==3.10
immutabledict==2.0.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
jax-cuda12-pjrt==0.4.30
jax-cuda12-plugin==0.4.30
Jinja2==3.1.4
jmp==0.0.4
joblib==1.4.2
keras==2.11.0
kiwisolver==1.4.7
libclang==18.1.1
lmdb==1.5.1
Markdown==3.7
MarkupSafe==3.0.2
ml-collections==0.1.0
ml_dtypes==0.5.0
MolVS==0.1.1
msgpack==1.1.0
multidict==6.1.0
multipledispatch==1.0.0
multiprocess==0.70.17
ndindex==1.9.2
networkx==3.2.1
numexpr==2.10.2
numpy==1.24.0
nvidia-cuda-nvcc-cu12==12.6.85
oauthlib==3.2.2
opt_einsum==3.4.0
packaging==24.2
pandas==2.2.3
pathos==0.3.3
pillow==11.0.0
pox==0.3.5
ppft==1.7.6.9
propcache==0.2.0
protobuf==3.19.6
psutil==6.1.0
py-cpuinfo==9.0.0
pyparsing==3.2.0
pyrr==0.10.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
rdkit-pypi==2022.9.4
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.5.3
scikit-learn==1.5.2
scipy==1.13.1
selfies==2.2.0
six==1.16.0
tables==3.9.2
tabulate==0.9.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
threadpoolctl==3.5.0
tokenizers==0.21.0
torch @ file:///home/xuyn/torch-1.12.0%2Bcu113-cp39-cp39-linux_x86_64.whl#sha256=566fbf0a8acc897f98f35992fc5d856d1b238652c2e6baf42a7e3f26c8e94ee4
torch-cluster @ file:///home/xuyn/torch_cluster-1.6.0%2Bpt112cu113-cp39-cp39-linux_x86_64.whl#sha256=14af0ee0b5d6a87b56c64ec22fdd0264e14f80f472cbc33aa47328077bdd0516
torch-geometric==2.6.1
torch-scatter @ file:///home/xuyn/torch_scatter-2.1.0%2Bpt112cu113-cp39-cp39-linux_x86_64.whl#sha256=ad2590451f514b1c67209c93a2723c522d90efc7c7c3476844dace5726b3094a
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
wrapt==1.17.0
yarl==1.18.0
zipp==3.21.0
